+++ client/src/app/site/pages/meetings/pages/motions/modules/change-recommendations/services/motion-diff.service/motion-diff.service.ts
@@ -1,22 +1,16 @@
 import { Injectable } from '@angular/core';
 import { TranslateService } from '@ngx-translate/core';
-import { djb2hash, splitStringKeepSeperator } from 'src/app/infrastructure/utils';
-import * as DomHelpers from 'src/app/infrastructure/utils/dom-helpers';
+import { djb2hash } from 'src/app/infrastructure/utils';
+import { replaceHtmlEntities } from 'src/app/infrastructure/utils/dom-helpers';
 
 import {
-    CommonAncestorData,
     DiffCache,
     DiffLinesInParagraph,
     ExtractedContent,
     LineRange
 } from '../../../../definitions';
-import { ViewUnifiedChangeType } from '../../definitions';
 import { ViewUnifiedChange } from '../../view-models';
-import { LineNumberedString, LineNumberingService, LineNumberRange } from '../line-numbering.service';
-
-const ELEMENT_NODE = Node.ELEMENT_NODE;
-const TEXT_NODE = Node.TEXT_NODE;
-const DOCUMENT_FRAGMENT_NODE = Node.DOCUMENT_FRAGMENT_NODE;
+import { LineNumberedString } from '../line-numbering.service';
 
 /**
  * Functionality regarding diffing, merging and extracting line ranges.
@@ -86,646 +80,11 @@ const DOCUMENT_FRAGMENT_NODE = Node.DOCUMENT_FRAGMENT_NODE;
     providedIn: `root`
 })
 export class MotionDiffService {
-    // @TODO Decide on a more sophisticated implementation
     private diffCache = new DiffCache();
 
     public constructor(
-        private lineNumberingService: LineNumberingService,
         private translate: TranslateService
     ) {}
-
-    /**
-     * Searches for the line breaking node within the given Document specified by the given lineNumber.
-     * This is performed by using a querySelector.
-     *
-     * @param {DocumentFragment} fragment
-     * @param {number} lineNumber
-     * @returns {Element}
-     */
-    public getLineNumberNode(fragment: DocumentFragment, lineNumber: number): Element | null {
-        return fragment?.querySelector(`os-linebreak.os-line-number.line-number-` + lineNumber);
-    }
-
-    /**
-     * This returns the first line breaking node within the given node.
-     * If none is found, `null` is returned.
-     *
-     * @param {Node} node
-     * @returns {Element}
-     */
-    private getFirstLineNumberNode(node: Node): Element | null {
-        return DomHelpers.getNodeByName(node, `OS-LINEBREAK`);
-    }
-
-    /**
-     * This returns the last line breaking node within the given node.
-     * If none is found, `null` is returned.
-     *
-     * @param {Node} node
-     * @returns {Element}
-     */
-    private getLastLineNumberNode(node: Node): Element | null {
-        return DomHelpers.getNodeByName(node, `OS-LINEBREAK`, true);
-    }
-
-    /**
-     * Returns a os linebreak element
-     * Example: <OS-LINEBREAK class="os-line-number line-number-23" data-line-number="23"/>
-     */
-    private getLineMarker(lineNumber: number, classes?: string): Node {
-        const lineMarker = document.createElement(`OS-LINEBREAK`);
-        lineMarker.setAttribute(`data-line-number`, lineNumber.toString(10));
-        lineMarker.setAttribute(`class`, classes ?? `os-line-number line-number-` + lineNumber.toString(10));
-
-        return lineMarker;
-    }
-
-    /**
-     * Adds elements like <OS-LINEBREAK class="os-line-number line-number-23" data-line-number="23"/>
-     * to a given fragment
-     *
-     * @param {DocumentFragment} fragment
-     */
-    public insertInternalLineMarkers(fragment: DocumentFragment): void {
-        if (fragment.querySelectorAll(`OS-LINEBREAK`).length > 0) {
-            // Prevent duplicate calls
-            return;
-        }
-        const lineNumbers = fragment.querySelectorAll(`span.os-line-number`);
-        let maxLineNumber = 0;
-
-        lineNumbers.forEach((insertBefore: Node) => {
-            const lineNumberElement = insertBefore as Element;
-            while (
-                insertBefore.parentNode?.nodeType !== DOCUMENT_FRAGMENT_NODE &&
-                DomHelpers.isFirstNonemptyChild(insertBefore.parentNode!, insertBefore)
-            ) {
-                insertBefore = insertBefore.parentNode!;
-            }
-            insertBefore.parentNode?.insertBefore(
-                this.getLineMarker(
-                    parseInt(lineNumberElement.getAttribute(`data-line-number`)),
-                    lineNumberElement.getAttribute(`class`)
-                ),
-                insertBefore
-            );
-            maxLineNumber = parseInt(lineNumberElement.getAttribute(`data-line-number`) as string, 10);
-        });
-
-        // Add one more "fake" line number at the end and beginning, so we can select the last line as well
-        fragment.appendChild(this.getLineMarker(maxLineNumber + 1));
-        fragment.insertBefore(this.getLineMarker(0), fragment.firstChild);
-    }
-
-    /**
-     * Returns information about the common ancestors of two given nodes.
-     *
-     * @param {Node} node1
-     * @param {Node} node2
-     * @returns {CommonAncestorData}
-     */
-    public getCommonAncestor(node1: Node, node2: Node): CommonAncestorData {
-        const trace1 = DomHelpers.getNodeContextTrace(node1);
-        const trace2 = DomHelpers.getNodeContextTrace(node2);
-        const childTrace1 = [];
-        const childTrace2 = [];
-        let commonAncestor: Node | null = null;
-        let commonIndex = 0;
-
-        for (let i = 0; i < trace1.length && i < trace2.length; i++) {
-            if (trace1[i] === trace2[i]) {
-                commonAncestor = trace1[i];
-                commonIndex = i;
-            }
-        }
-        for (let i = commonIndex + 1; i < trace1.length; i++) {
-            childTrace1.push(trace1[i]);
-        }
-        for (let i = commonIndex + 1; i < trace2.length; i++) {
-            childTrace2.push(trace2[i]);
-        }
-        return {
-            commonAncestor: commonAncestor!,
-            trace1: childTrace1,
-            trace2: childTrace2,
-            index: commonIndex
-        };
-    }
-
-    /**
-     * This converts a HTML Node element into a rendered HTML string.
-     *
-     * @param {Node} node
-     * @returns {string}
-     */
-    private serializeTag(node: Node): string {
-        if (node.nodeType !== ELEMENT_NODE) {
-            // Fragments are only placeholders and do not have an HTML representation
-            return ``;
-        }
-
-        return DomHelpers.serializeTag(node);
-    }
-
-    /**
-     * This performs HTML normalization to prevent the Diff-Algorithm from detecting changes when there are actually
-     * none. Common problems covered by this method are differently ordered Attributes of HTML elements or HTML-encoded
-     * special characters.
-     *
-     * @param {string} html
-     * @returns {string}
-     * @private
-     */
-    public normalizeHtmlForDiff(html: string): string {
-        html = DomHelpers.sortHtmlAttributes(html);
-        html = DomHelpers.normalizeStyleAttributes(html);
-        html = DomHelpers.htmlToUppercase(html);
-
-        // remove whitespaces infront of closing tags
-        html = html
-            .replace(/\s+<\/P>/gi, `</P>`)
-            .replace(/\s+<\/DIV>/gi, `</DIV>`)
-            .replace(/\s+<\/LI>/gi, `</LI>`);
-        html = html.replace(/\s+<LI>/gi, `<LI>`).replace(/<\/LI>\s+/gi, `</LI>`);
-
-        html = html.replace(/\u00A0/g, ` `); // replace no break space
-        html = html.replace(/\u2013/g, `-`);
-        html = DomHelpers.replaceHtmlEntities(html);
-
-        // Newline characters: after closing block-level-elements, but not after BR (which is inline)
-        html = html.replace(/(<br *\/?>)\n/gi, `$1`);
-        html = html.replace(/[ \n\t]+/gi, ` `);
-        html = html.replace(/(<\/(div|p|ul|li|blockquote>)>) /gi, `$1\n`);
-
-        return html;
-    }
-
-    /**
-     * Adapted from http://ejohn.org/projects/javascript-diff-algorithm/
-     * by John Resig, MIT License
-     *
-     * @param {array} oldArr
-     * @param {array} newArr
-     * @returns {object}
-     */
-    private diffArrays(oldArr: any, newArr: any): any {
-        const ns: any = {};
-        const os: any = {};
-
-        for (let i = 0; i < newArr.length; i++) {
-            if (ns[newArr[i]] === undefined) {
-                ns[newArr[i]] = { rows: [], o: null };
-            }
-            ns[newArr[i]].rows.push(i);
-        }
-
-        for (let i = 0; i < oldArr.length; i++) {
-            if (os[oldArr[i]] === undefined) {
-                os[oldArr[i]] = { rows: [], n: null };
-            }
-            os[oldArr[i]].rows.push(i);
-        }
-
-        for (const i in ns) {
-            if (ns[i].rows.length === 1 && typeof os[i] !== `undefined` && os[i].rows.length === 1) {
-                newArr[ns[i].rows[0]] = { text: newArr[ns[i].rows[0]], row: os[i].rows[0] };
-                oldArr[os[i].rows[0]] = { text: oldArr[os[i].rows[0]], row: ns[i].rows[0] };
-            } else if (
-                ns[i].rows.length >= 1 &&
-                ns[i].rows.indexOf(0) !== -1 &&
-                os[i] !== undefined &&
-                os[i].rows.indexOf(0) !== -1
-            ) {
-                newArr[0] = { text: newArr[0], row: 0 };
-                oldArr[0] = { text: oldArr[0], row: 0 };
-            }
-        }
-
-        for (let i = 0; i < newArr.length - 1; i++) {
-            if (
-                newArr[i].text !== null &&
-                newArr[i + 1].text === undefined &&
-                newArr[i].row + 1 < oldArr.length &&
-                oldArr[newArr[i].row + 1].text === undefined &&
-                newArr[i + 1] === oldArr[newArr[i].row + 1]
-            ) {
-                newArr[i + 1] = { text: newArr[i + 1], row: newArr[i].row + 1 };
-                oldArr[newArr[i].row + 1] = { text: oldArr[newArr[i].row + 1], row: i + 1 };
-            }
-        }
-
-        for (let i = newArr.length - 1; i > 0; i--) {
-            if (
-                newArr[i].text !== null &&
-                newArr[i - 1].text === undefined &&
-                newArr[i].row > 0 &&
-                oldArr[newArr[i].row - 1].text === undefined &&
-                newArr[i - 1] === oldArr[newArr[i].row - 1]
-            ) {
-                newArr[i - 1] = { text: newArr[i - 1], row: newArr[i].row - 1 };
-                oldArr[newArr[i].row - 1] = { text: oldArr[newArr[i].row - 1], row: i - 1 };
-            }
-        }
-
-        // This fixes the problem tested by "does not lose words when changes are moved X-wise"
-        let lastRow = 0;
-        for (let z = 0; z < newArr.length; z++) {
-            if (newArr[z].row && newArr[z].row > lastRow) {
-                lastRow = newArr[z].row;
-            }
-            if (newArr[z].row && newArr[z].row < lastRow) {
-                oldArr[newArr[z].row] = oldArr[newArr[z].row].text;
-                newArr[z] = newArr[z].text;
-            }
-        }
-
-        return { o: oldArr, n: newArr };
-    }
-
-    /**
-     * This method splits a string into an array of strings, such as that it can be used by the diff method.
-     * Mainly it tries to split it into single words, but prevents HTML tags from being split into different elements.
-     *
-     * @param {string} str
-     * @returns {string[]}
-     */
-    private tokenizeHtml(str: string): string[] {
-        let res = [str];
-        for (const splitConf of [
-            { by: `<`, type: `prepend` },
-            { by: `>`, type: `append` },
-            { by: ` ` },
-            { by: `.` },
-            { by: `,` },
-            { by: `!` },
-            { by: `-` },
-            { by: `\n`, type: `append` }
-        ]) {
-            const newArr = [];
-            for (const str of res) {
-                // Don't split HTML tags
-                if (str[0] === `<` && splitConf.by !== `<` && splitConf.by !== `>`) {
-                    newArr.push(str);
-                    continue;
-                }
-
-                newArr.push(...splitStringKeepSeperator(str, splitConf.by, splitConf.type));
-            }
-            res = newArr;
-        }
-
-        return res.filter(el => el !== ``);
-    }
-
-    /**
-     * Given two strings, this method generates a consolidated new string that indicates the operations necessary
-     * to get from `oldStr` to `newStr` by <ins>...</ins> and <del>...</del>-Tags
-     *
-     * @param {string} oldStr
-     * @param {string} newStr
-     * @returns {string}
-     */
-    private diffString(oldStr: string, newStr: string): string {
-        oldStr = this.normalizeHtmlForDiff(oldStr.replace(/\s+$/, ``).replace(/^\s+/, ``));
-        newStr = this.normalizeHtmlForDiff(newStr.replace(/\s+$/, ``).replace(/^\s+/, ``));
-
-        const out = this.diffArrays(this.tokenizeHtml(oldStr), this.tokenizeHtml(newStr));
-
-        let str = ``;
-        if (out.n.length === 0) {
-            for (const o of out.o) {
-                str += `<del>` + o + `</del>`;
-            }
-        } else {
-            if (out.n[0].text === undefined) {
-                for (let k = 0; k < out.o.length && out.o[k].text === undefined; k++) {
-                    str += `<del>` + out.o[k] + `</del>`;
-                }
-            }
-
-            let currOldRow = 0;
-            for (let i = 0; i < out.n.length; i++) {
-                if (out.n[i].text === undefined) {
-                    if (out.n[i] !== ``) {
-                        str += `<ins>` + out.n[i] + `</ins>`;
-                    }
-                } else if (out.n[i].row < currOldRow) {
-                    str += `<ins>` + out.n[i].text + `</ins>`;
-                } else {
-                    let pre = ``;
-
-                    if (i + 1 < out.n.length && out.n[i + 1].row !== undefined && out.n[i + 1].row > out.n[i].row + 1) {
-                        for (let n = out.n[i].row + 1; n < out.n[i + 1].row; n++) {
-                            if (out.o[n].text === undefined) {
-                                pre += `<del>` + out.o[n] + `</del>`;
-                            } else {
-                                pre += `<del>` + out.o[n].text + `</del>`;
-                            }
-                        }
-                    } else {
-                        for (let j = out.n[i].row + 1; j < out.o.length && out.o[j].text === undefined; j++) {
-                            pre += `<del>` + out.o[j] + `</del>`;
-                        }
-                    }
-                    str += out.n[i].text + pre;
-
-                    currOldRow = out.n[i].row;
-                }
-            }
-        }
-
-        return str.replace(/^\s+/g, ``).replace(/\s+$/g, ``).replace(/ {2,}/g, ` `);
-    }
-
-    /**
-     * This detects if a given string contains broken HTML. This can happen when the Diff accidentally produces
-     * wrongly nested HTML tags.
-     *
-     * @param {string} html
-     * @returns {boolean}
-     * @private
-     */
-    private diffDetectBrokenDiffHtml(html: string): boolean {
-        // If other HTML tags are contained within INS/DEL (e.g. "<ins>Test</p></ins>"), let's better be cautious
-        // The "!!(found=...)"-construction is only used to make jshint happy :)
-        const findDel = /<del>([\s\S]*?)<\/del>/gi;
-        const findIns = /<ins>([\s\S]*?)<\/ins>/gi;
-        let found: RegExpExecArray;
-        let inner: string;
-        while ((found = findDel.exec(html))) {
-            inner = found[1].replace(/<br[^>]*>/gi, ``);
-            if (!DomHelpers.isValidInlineHtml(inner)) {
-                return true;
-            }
-        }
-        while ((found = findIns.exec(html))) {
-            inner = found[1].replace(/<br[^>]*>/gi, ``);
-            if (!DomHelpers.isValidInlineHtml(inner)) {
-                return true;
-            }
-        }
-
-        // If non of the conditions up to now is met, we consider the diff as being sane
-        return false;
-    }
-
-    /**
-     * This function removes color-Attributes from the styles of this node or a descendant,
-     * as they interfer with the green/red color in HTML and PDF
-     *
-     * For the moment, it is sufficient to do this only in paragraph diff mode, as we fall back to this mode anyway
-     * once we encounter SPANs or other tags inside of INS/DEL-tags
-     *
-     * @param {Element} node
-     * @private
-     */
-    private removeColorStyles(node: Element): void {
-        const styles = node.getAttribute(`style`);
-        if (styles && styles.indexOf(`color`) > -1) {
-            const stylesNew: any[] = [];
-            styles.split(`;`).forEach((style: string): void => {
-                if (!style.match(/^\s*color\s*:/i)) {
-                    stylesNew.push(style);
-                }
-            });
-            if (stylesNew.join(`;`) === ``) {
-                node.removeAttribute(`style`);
-            } else {
-                node.setAttribute(`style`, stylesNew.join(`;`));
-            }
-        }
-        for (const child of node.childNodes) {
-            if (child.nodeType === ELEMENT_NODE) {
-                this.removeColorStyles((child as Element));
-            }
-        }
-    }
-
-    /**
-     * This fixes a very specific, really weird bug that is tested in the test case "does not a change in a very
-     * specific case.
-     *
-     * @param {string}diffStr
-     * @return {string}
-     */
-    private fixWrongChangeDetection(diffStr: string): string {
-        if (diffStr.indexOf(`<del>`) === -1 || diffStr.indexOf(`<ins>`) === -1) {
-            return diffStr;
-        }
-
-        const findDelGroupFinder = /(?:<del>.*?<\/del>)+/gi;
-        let found: RegExpExecArray;
-        let returnStr = diffStr;
-
-        while ((found = findDelGroupFinder.exec(diffStr))) {
-            const del = found[0];
-            const split = returnStr.split(del);
-
-            const findsGroupFinder = /^(?:<ins>.*?<\/ins>)+/gi;
-            const foundIns = findsGroupFinder.exec(split[1]);
-            if (foundIns) {
-                const ins = foundIns[0];
-
-                let delShortened = del
-                    .replace(
-                        /<del>((<BR CLASS="os-line-break"><\/del><del>)?(<span[^>]+os-line-number[^>]+?>)(\s|<\/?del>)*<\/span>)<\/del>/gi,
-                        ``
-                    )
-                    .replace(/<\/del><del>/g, ``);
-                const insConv = ins
-                    .replace(/<ins>/g, `<del>`)
-                    .replace(/<\/ins>/g, `</del>`)
-                    .replace(/<\/del><del>/g, ``);
-                if (delShortened.indexOf(insConv) !== -1) {
-                    delShortened = delShortened.replace(insConv, ``);
-                    if (delShortened === ``) {
-                        returnStr = returnStr.replace(del + ins, del.replace(/<del>/g, ``).replace(/<\/del>/g, ``));
-                    }
-                }
-            }
-        }
-        return returnStr;
-    }
-
-    /**
-     * Converts a given HTML node into HTML string and optionally strips line number nodes from it.
-     *
-     * @param {Node} node
-     * @param {boolean} stripLineNumbers
-     * @returns {string}
-     */
-    private serializeDom(node: Node, stripLineNumbers: boolean): string {
-        if (node.nodeType === TEXT_NODE) {
-            return node.nodeValue!.replace(/</g, `&lt;`).replace(/>/g, `&gt;`);
-        }
-        if (
-            stripLineNumbers &&
-            (this.lineNumberingService.isOsLineNumberNode(node) || this.lineNumberingService.isOsLineBreakNode(node))
-        ) {
-            return ``;
-        }
-        if (node.nodeName === `OS-LINEBREAK`) {
-            return ``;
-        }
-        if (node.nodeName === `BR`) {
-            const element = node as Element;
-            let br = `<BR`;
-            for (const attibutes of element.attributes) {
-                const attr = attibutes;
-                br += ` ` + attr.name + `="` + attr.value + `"`;
-            }
-            return br + `>`;
-        }
-
-        let html = this.serializeTag(node);
-        for (const child of node.childNodes) {
-            if (child.nodeType === TEXT_NODE) {
-                html += child
-                    .nodeValue!.replace(/&/g, `&amp;`)
-                    .replace(/</g, `&lt;`)
-                    .replace(/>/g, `&gt;`);
-            } else {
-                html += this.serializeDom(child, stripLineNumbers);
-            }
-        }
-        if (node.nodeType !== DOCUMENT_FRAGMENT_NODE) {
-            html += `</` + node.nodeName + `>`;
-        }
-
-        return html;
-    }
-
-    /**
-     * When a <li> with a os-split-before-class (set by extractRangeByLineNumbers) is edited when creating a
-     * change recommendation and is split again in CKEditor, the second list items also gets that class.
-     * This is not correct however, as the second one actually is a new list item. So we need to remove it again.
-     *
-     * @param {string} html
-     * @returns {string}
-     */
-    public removeDuplicateClassesInsertedByCkeditor(html: string): string {
-        const fragment = DomHelpers.htmlToFragment(html);
-        const items = fragment.querySelectorAll(`li.os-split-before`);
-        for (const item of items) {
-            if (!DomHelpers.isFirstNonemptyChild(item.parentNode!, item)) {
-                DomHelpers.removeCSSClass(item, `os-split-before`);
-            }
-        }
-        return this.serializeDom(fragment, false);
-    }
-
-    /**
-     * Given a DOM tree and a specific node within that tree, this method returns the HTML string from the beginning
-     * of this tree up to this node.
-     * The returned string in itself is not renderable, as it stops in the middle of the complete HTML, with
-     * opened tags.
-     *
-     * Implementation hint: the first element of "toChildTrace" array needs to be a child element of "node"
-     * @param {Node} node
-     * @param {Node[]} toChildTrace
-     * @param {boolean} stripLineNumbers
-     * @returns {string}
-     */
-    public serializePartialDomToChild(node: Node, toChildTrace: Node[], stripLineNumbers: boolean): string {
-        if (toChildTrace.length === 0) {
-            return ``;
-        }
-        if (this.lineNumberingService.isOsLineNumberNode(node) || this.lineNumberingService.isOsLineBreakNode(node)) {
-            return ``;
-        }
-        if (node.nodeName === `OS-LINEBREAK`) {
-            return ``;
-        }
-
-        let html = this.serializeTag(node);
-        let found = false;
-
-        for (let i = 0; i < node.childNodes.length && !found; i++) {
-            if (node.childNodes[i] === toChildTrace[0]) {
-                found = true;
-                const childElement = node.childNodes[i] as Element;
-                const remainingTrace = toChildTrace;
-                remainingTrace.shift();
-                if (!this.lineNumberingService.isOsLineNumberNode(childElement) && remainingTrace.length > 0) {
-                    html += this.serializePartialDomToChild(childElement, remainingTrace, stripLineNumbers);
-                }
-            } else if (node.childNodes[i].nodeType === TEXT_NODE) {
-                html += node.childNodes[i].nodeValue;
-            } else {
-                const childElement = node.childNodes[i] as Element;
-                if (
-                    !stripLineNumbers ||
-                    (!this.lineNumberingService.isOsLineNumberNode(childElement) &&
-                        !this.lineNumberingService.isOsLineBreakNode(childElement))
-                ) {
-                    html += this.serializeDom(childElement, stripLineNumbers);
-                }
-            }
-        }
-        if (!found) {
-            throw new Error(`Inconsistency or invalid call of this function detected (to)`);
-        }
-        return html;
-    }
-
-    /**
-     * Given a DOM tree and a specific node within that tree, this method returns the HTML string beginning after this
-     * node to the end of the tree.
-     * The returned string in itself is not renderable, as it starts in the middle of the complete HTML
-     * with opened tags.
-     *
-     * Implementation hint: the first element of "fromChildTrace" array needs to be a child element of "node"
-     * @param {Node} node
-     * @param {Node[]} fromChildTrace
-     * @param {boolean} stripLineNumbers
-     * @returns {string}
-     */
-    public serializePartialDomFromChild(node: Node, fromChildTrace: Node[], stripLineNumbers: boolean): string {
-        if (fromChildTrace.length === 0) {
-            return ``;
-        }
-        if (this.lineNumberingService.isOsLineNumberNode(node) || this.lineNumberingService.isOsLineBreakNode(node)) {
-            return ``;
-        }
-        if (node.nodeName === `OS-LINEBREAK`) {
-            return ``;
-        }
-
-        let html = ``;
-        let found = false;
-        for (const child of node.childNodes) {
-            if (child === fromChildTrace[0]) {
-                found = true;
-                const childElement = child as Element;
-                const remainingTrace = fromChildTrace;
-                remainingTrace.shift();
-                if (!this.lineNumberingService.isOsLineNumberNode(childElement) && remainingTrace.length > 0) {
-                    html += this.serializePartialDomFromChild(childElement, remainingTrace, stripLineNumbers);
-                }
-            } else if (found) {
-                if (child.nodeType === TEXT_NODE) {
-                    html += child.nodeValue;
-                } else {
-                    const childElement = child as Element;
-                    if (
-                        !stripLineNumbers ||
-                        (!this.lineNumberingService.isOsLineNumberNode(childElement) &&
-                            !this.lineNumberingService.isOsLineBreakNode(childElement))
-                    ) {
-                        html += this.serializeDom(childElement, stripLineNumbers);
-                    }
-                }
-            }
-        }
-        if (!found) {
-            throw new Error(`Inconsistency or invalid call of this function detected (from)`);
-        }
-        if (node.nodeType !== DOCUMENT_FRAGMENT_NODE) {
-            html += `</` + node.nodeName + `>`;
-        }
-
-        return html;
-    }
 
     /**
      * Returns the HTML snippet between two given line numbers.
@@ -771,170 +130,9 @@ export class MotionDiffService {
             return cached;
         }
 
-        const fragment = DomHelpers.htmlToFragment(html);
-        this.insertInternalLineMarkers(fragment);
+        // TODO
 
-        let toLineNumber: number;
-        if (toLine === null) {
-            const internalLineMarkers = fragment.querySelectorAll(`OS-LINEBREAK`);
-            const lastMarker = internalLineMarkers[internalLineMarkers.length - 1] as Element;
-            toLineNumber = parseInt(lastMarker.getAttribute(`data-line-number`) as string, 10);
-        } else {
-            toLineNumber = toLine + 1;
-        }
-
-        const fromLineNumberNode = this.getLineNumberNode(fragment, fromLine);
-        const toLineNumberNode = toLineNumber ? this.getLineNumberNode(fragment, toLineNumber) : null;
-        const ancestorData = this.getCommonAncestor(fromLineNumberNode as Element, toLineNumberNode as Element);
-
-        const fromChildTraceRel = ancestorData.trace1;
-        const fromChildTraceAbs = DomHelpers.getNodeContextTrace(fromLineNumberNode as Element);
-        const toChildTraceRel = ancestorData.trace2;
-        const toChildTraceAbs = DomHelpers.getNodeContextTrace(toLineNumberNode as Element);
-        const ancestor = ancestorData.commonAncestor;
-        let htmlOut = ``;
-        let outerContextStart = ``;
-        let outerContextEnd = ``;
-        let innerContextStart = ``;
-        let innerContextEnd = ``;
-        let previousHtmlEndSnippet = ``;
-        let followingHtmlStartSnippet = ``;
-
-        fromChildTraceAbs.shift();
-        const previousHtml = this.serializePartialDomToChild(fragment, fromChildTraceAbs, false);
-
-        toChildTraceAbs.shift();
-        const followingHtml = this.serializePartialDomFromChild(fragment, toChildTraceAbs, false);
-
-        let currNode: Node = fromLineNumberNode as Element;
-        let isSplit = false;
-        while (currNode.parentNode) {
-            if (!DomHelpers.isFirstNonemptyChild(currNode.parentNode, currNode)) {
-                isSplit = true;
-            }
-            if (isSplit) {
-                DomHelpers.addCSSClass(currNode.parentNode, `os-split-before`);
-            }
-            if (currNode.nodeName !== `OS-LINEBREAK`) {
-                previousHtmlEndSnippet += `</` + currNode.nodeName + `>`;
-            }
-            currNode = currNode.parentNode;
-        }
-
-        currNode = toLineNumberNode as Element;
-        isSplit = false;
-        while (currNode.parentNode) {
-            if (!DomHelpers.isFirstNonemptyChild(currNode.parentNode, currNode)) {
-                isSplit = true;
-            }
-            if (isSplit) {
-                DomHelpers.addCSSClass(currNode.parentNode, `os-split-after`);
-            }
-            if (currNode.parentNode.nodeName === `OL`) {
-                const parentElement = currNode.parentNode as Element;
-                const fakeOl = parentElement.cloneNode(false) as any;
-                const offset = parentElement.getAttribute(`start`)
-                    ? parseInt(parentElement.getAttribute(`start`) as string, 10) - 1
-                    : 0;
-                fakeOl.setAttribute(
-                    `start`,
-                    (
-                        (DomHelpers.getNthOfListItem(parentElement, toLineNumberNode as Element) as number) + offset
-                    ).toString()
-                );
-                followingHtmlStartSnippet = this.serializeTag(fakeOl) + followingHtmlStartSnippet;
-            } else {
-                followingHtmlStartSnippet = this.serializeTag(currNode.parentNode) + followingHtmlStartSnippet;
-            }
-            currNode = currNode.parentNode;
-        }
-
-        isSplit = false;
-        for (let i = 0, found = false; i < fromChildTraceRel.length && !found; i++) {
-            if (fromChildTraceRel[i].nodeName === `OS-LINEBREAK`) {
-                found = true;
-            } else {
-                if (!DomHelpers.isFirstNonemptyChild(fromChildTraceRel[i], fromChildTraceRel[i + 1])) {
-                    isSplit = true;
-                }
-                if (fromChildTraceRel[i].nodeName === `OL`) {
-                    const element = fromChildTraceRel[i] as Element;
-                    const fakeOl = element.cloneNode(false) as any;
-                    const offset = element.getAttribute(`start`)
-                        ? parseInt(element.getAttribute(`start`) as string, 10) - 1
-                        : 0;
-                    fakeOl.setAttribute(
-                        `start`,
-                        (
-                            offset + (DomHelpers.getNthOfListItem(element, fromLineNumberNode as Element) as number)
-                        ).toString()
-                    );
-                    innerContextStart += this.serializeTag(fakeOl);
-                } else {
-                    if (i < fromChildTraceRel.length - 1 && isSplit) {
-                        DomHelpers.addCSSClass(fromChildTraceRel[i], `os-split-before`);
-                    }
-                    innerContextStart += this.serializeTag(fromChildTraceRel[i]);
-                }
-            }
-        }
-        for (let i = 0, found = false; i < toChildTraceRel.length && !found; i++) {
-            if (toChildTraceRel[i].nodeName === `OS-LINEBREAK`) {
-                found = true;
-            } else {
-                innerContextEnd = `</` + toChildTraceRel[i].nodeName + `>` + innerContextEnd;
-            }
-        }
-
-        for (let i = 0, found = false; i < ancestor.childNodes.length; i++) {
-            if (ancestor.childNodes[i] === fromChildTraceRel[0]) {
-                found = true;
-                fromChildTraceRel.shift();
-                htmlOut += this.serializePartialDomFromChild(ancestor.childNodes[i], fromChildTraceRel, true);
-            } else if (ancestor.childNodes[i] === toChildTraceRel[0]) {
-                found = false;
-                toChildTraceRel.shift();
-                htmlOut += this.serializePartialDomToChild(ancestor.childNodes[i], toChildTraceRel, true);
-            } else if (found === true) {
-                htmlOut += this.serializeDom(ancestor.childNodes[i], true);
-            }
-        }
-
-        currNode = ancestor;
-        while (currNode.parentNode) {
-            if (currNode.nodeName === `OL`) {
-                const currElement = currNode as Element;
-                const fakeOl = currElement.cloneNode(false) as any;
-                const offset = currElement.getAttribute(`start`)
-                    ? parseInt(currElement.getAttribute(`start`) as string, 10) - 1
-                    : 0;
-                fakeOl.setAttribute(
-                    `start`,
-                    ((DomHelpers.getNthOfListItem(currElement, fromLineNumberNode as Element) as any) + offset).toString()
-                );
-                outerContextStart = this.serializeTag(fakeOl) + outerContextStart;
-            } else {
-                outerContextStart = this.serializeTag(currNode) + outerContextStart;
-            }
-            outerContextEnd += `</` + currNode.nodeName + `>`;
-            currNode = currNode.parentNode;
-        }
-
-        const ret = {
-            html: htmlOut,
-            ancestor,
-            outerContextStart,
-            outerContextEnd,
-            innerContextStart,
-            innerContextEnd,
-            previousHtml,
-            previousHtmlEndSnippet,
-            followingHtml,
-            followingHtmlStartSnippet
-        };
-
-        this.diffCache.put(cacheKey, ret);
-        return ret;
+        return null;
     }
 
     /**
@@ -950,111 +148,6 @@ export class MotionDiffService {
     }
 
     /**
-     * Convenience method that takes the html-attribute from an extractRangeByLineNumbers()-method,
-     * wraps it with the context and adds line numbers.
-     *
-     * @param {ExtractedContent} diff
-     * @param {number} lineLength
-     * @param {number} firstLine
-     */
-    private formatDiffWithLineNumbers(diff: ExtractedContent, lineLength: number, firstLine: number): string {
-        let text = this.formatDiff(diff);
-        text = this.lineNumberingService.insertLineNumbers({ html: text, lineLength, firstLine });
-        return text;
-    }
-
-    /**
-     * This is a workardoun to prevent the last word of the inserted text from accidently being merged with the
-     * first word of the following line.
-     *
-     * This happens as trailing spaces in the change recommendation's text are frequently stripped,
-     * which is pretty nasty if the original text goes on after the affected line. So we insert a space
-     * if the original line ends with one.
-     *
-     * @param {Element|DocumentFragment} element
-     */
-    private insertDanglingSpace(element: Element | DocumentFragment): void {
-        if (element.childNodes.length > 0) {
-            let lastChild = element.childNodes[element.childNodes.length - 1];
-            if (
-                lastChild.nodeType === TEXT_NODE &&
-                !lastChild.nodeValue!.match(/[\S]/) &&
-                element.childNodes.length > 1
-            ) {
-                // If the text node only contains whitespaces, chances are high it's just space between block elmeents,
-                // like a line break between </LI> and </UL>
-                lastChild = element.childNodes[element.childNodes.length - 2];
-            }
-            if (lastChild.nodeType === TEXT_NODE) {
-                if (lastChild.nodeValue === `` || lastChild.nodeValue!.substr(-1) !== ` `) {
-                    lastChild.nodeValue += ` `;
-                }
-            } else {
-                this.insertDanglingSpace((lastChild as Element));
-            }
-        }
-    }
-
-    /**
-     * This functions merges two arrays of nodes. The last element of nodes1 and the first element of nodes2
-     * are merged, if they are of the same type.
-     *
-     * This is done recursively until a TEMPLATE-Tag is found, which was inserted in this.replaceLines.
-     * Using a TEMPLATE-Tag is a rather dirty hack, as it is allowed inside any other element, including <ul>.
-     *
-     * @param {Node[]} nodes1
-     * @param {Node[]} nodes2
-     * @returns {Node[]}
-     */
-    public replaceLinesMergeNodeArrays(nodes1: Node[], nodes2: Node[]): Node[] {
-        if (nodes1.length === 0 || nodes2.length === 0) {
-            return nodes1.length ? nodes1 : nodes2;
-        }
-
-        const out: Node[] = nodes1.slice(0, -1);
-        const lastNode: Node = nodes1[nodes1.length - 1];
-        const firstNode: Node = nodes2[0];
-        if (lastNode.nodeType === TEXT_NODE && firstNode.nodeType === TEXT_NODE) {
-            const newTextNode: Text = lastNode.ownerDocument!.createTextNode(lastNode.nodeValue! + firstNode.nodeValue);
-            out.push(newTextNode);
-        } else if (lastNode.nodeName === firstNode.nodeName) {
-            const lastElement: Element = lastNode as Element;
-            const newNode: HTMLElement = lastNode.ownerDocument!.createElement(lastNode.nodeName);
-
-            for (const attr of Array.from(lastElement.attributes)) {
-                newNode.setAttribute(attr.name, attr.value);
-            }
-
-            // Remove #text nodes inside of List elements (OL/UL), as they are confusing
-            let lastChildren: Node[];
-            let firstChildren: Node[];
-            if (lastElement.nodeName === `OL` || lastElement.nodeName === `UL`) {
-                lastChildren = Array.from(lastElement.childNodes).filter(child => child.nodeType === ELEMENT_NODE);
-                firstChildren = Array.from(firstNode.childNodes).filter(child => child.nodeType === ELEMENT_NODE);
-            } else {
-                lastChildren = Array.from(lastElement.childNodes);
-                firstChildren = Array.from(firstNode.childNodes);
-            }
-
-            const children = this.replaceLinesMergeNodeArrays(lastChildren, firstChildren) as Node[];
-            for (const child of children) {
-                newNode.appendChild(child);
-            }
-
-            out.push(newNode);
-        } else {
-            if (lastNode.nodeName !== `TEMPLATE`) {
-                out.push(lastNode);
-            }
-            if (firstNode.nodeName !== `TEMPLATE`) {
-                out.push(firstNode);
-            }
-        }
-
-        return out.concat(nodes2.slice(1, nodes2.length));
-    }
-
-    /**
      * This returns the line number range in which changes (insertions, deletions) are encountered.
      * As in extractRangeByLineNumbers(), "to" refers to the line breaking element at the end, i.e. the start of the
      * following line.
@@ -1065,50 +158,7 @@ export class MotionDiffService {
      * @returns {LineRange}
      */
     public detectAffectedLineRange(diffHtml: string): LineRange | null {
-        const cacheKey = djb2hash(diffHtml);
-        const cached = this.diffCache.get(cacheKey);
-        if (cached) {
-            return cached;
-        }
-
-        const fragment = DomHelpers.htmlToFragment(diffHtml);
-
-        this.insertInternalLineMarkers(fragment);
-
-        const changes = fragment.querySelectorAll(`ins, del, .insert, .delete`);
-        const firstChange = changes.item(0);
-        const lastChange = changes.item(changes.length - 1);
-
-        if (!firstChange || !lastChange) {
-            // There are no changes
-            return null;
-        }
-
-        const firstTrace = DomHelpers.getNodeContextTrace(firstChange);
-        let lastLineNumberBefore = null;
-        for (let j = firstTrace.length - 1; j >= 0 && lastLineNumberBefore === null; j--) {
-            const prevSiblings = DomHelpers.getAllPrevSiblingsReversed(firstTrace[j]);
-            for (let i = 0; i < prevSiblings.length && lastLineNumberBefore === null; i++) {
-                lastLineNumberBefore = this.getLastLineNumberNode(prevSiblings[i]);
-            }
-        }
-
-        const lastTrace = DomHelpers.getNodeContextTrace(lastChange);
-        let firstLineNumberAfter = null;
-        for (let j = lastTrace.length - 1; j >= 0 && firstLineNumberAfter === null; j--) {
-            const nextSiblings = DomHelpers.getAllNextSiblings(lastTrace[j]);
-            for (let i = 0; i < nextSiblings.length && firstLineNumberAfter === null; i++) {
-                firstLineNumberAfter = this.getFirstLineNumberNode(nextSiblings[i]);
-            }
-        }
-
-        const range = {
-            from: parseInt(lastLineNumberBefore!.getAttribute(`data-line-number`) as string, 10),
-            to: parseInt(firstLineNumberAfter!.getAttribute(`data-line-number`) as string, 10) - 1
-        };
-
-        this.diffCache.put(cacheKey, range);
-        return range;
+        return null;
     }
 
     /**
@@ -1119,29 +169,7 @@ export class MotionDiffService {
      * @returns {string}
      */
     public diffHtmlToFinalText(html: string): string {
-        const fragment = DomHelpers.htmlToFragment(html);
-
-        const delNodes = fragment.querySelectorAll(`.delete, del`);
-        for (const del of delNodes) {
-            del.parentNode!.removeChild(del);
-        }
-
-        const insNodes = fragment.querySelectorAll(`ins`);
-        for (const ins of insNodes) {
-            while (ins.childNodes.length > 0) {
-                const child = ins.childNodes.item(0);
-                ins.removeChild(child);
-                ins.parentNode!.insertBefore(child, ins);
-            }
-            ins.parentNode!.removeChild(ins);
-        }
-
-        const insertNodes = fragment.querySelectorAll(`.insert`);
-        for (const insert of insertNodes) {
-            DomHelpers.removeCSSClass(insert, `insert`);
-        }
-
-        return this.serializeDom(fragment, false);
+        return null;
     }
 
     /**
@@ -1158,128 +186,7 @@ export class MotionDiffService {
      * @param {number} toLine
      */
     public replaceLines(oldHtml: string, newHTML: string, fromLine: number, toLine: number): string {
-        const data = this.extractRangeByLineNumbers(oldHtml, fromLine, toLine);
-        const previousHtml = data.previousHtml + `<TEMPLATE></TEMPLATE>` + data.previousHtmlEndSnippet;
-        const previousFragment = DomHelpers.htmlToFragment(previousHtml);
-        const followingHtml = data.followingHtmlStartSnippet + `<TEMPLATE></TEMPLATE>` + data.followingHtml;
-        const followingFragment = DomHelpers.htmlToFragment(followingHtml);
-        const newFragment = DomHelpers.htmlToFragment(newHTML);
-
-        if (data.html.length > 0 && data.html.substr(-1) === ` `) {
-            this.insertDanglingSpace(newFragment);
-        }
-
-        let merged = this.replaceLinesMergeNodeArrays(
-            Array.prototype.slice.call(previousFragment.childNodes),
-            Array.prototype.slice.call(newFragment.childNodes)
-        );
-        merged = this.replaceLinesMergeNodeArrays(merged, Array.prototype.slice.call(followingFragment.childNodes));
-
-        const mergedFragment = document.createDocumentFragment();
-        for (const merge of merged) {
-            mergedFragment.appendChild(merge);
-        }
-
-        const forgottenTemplates = mergedFragment.querySelectorAll(`TEMPLATE`);
-        for (const forgottenTemp of forgottenTemplates) {
-            const el = forgottenTemp;
-            el.parentNode!.removeChild(el);
-        }
-
-        const forgottenSplitClasses = mergedFragment.querySelectorAll(`.os-split-before, .os-split-after`);
-        for (const forgottenSplit of forgottenSplitClasses) {
-            DomHelpers.removeCSSClass(forgottenSplit, `os-split-before`);
-            DomHelpers.removeCSSClass(forgottenSplit, `os-split-after`);
-        }
-
-        return this.serializeDom(mergedFragment, true);
-    }
-
-    private removeLines(oldHtml: string, fromLine: number, toLine: number): string {
-        return this.replaceLines(oldHtml, ``, fromLine, toLine);
-    }
-
-    /**
-     * Hint: as replaceLines does not work with specific points in the text anymore, but full lines, inserting
-     * only works by using the workaround of selecting a "negative line", which results in no removal.
-     *
-     * Only mind that the inserted HTML needs to be wrapped in similar tags to the preceding text; that is, if the
-     * previous text is within a UL/LI construct and insertedHtml is supposted to be inserted within that LI,
-     * it needs to be wrapped accordingly.
-     */
-    private insertLines(oldHtml: string, atLineNumber: number, insertedHtml: string): string {
-        return this.replaceLines(oldHtml, insertedHtml, atLineNumber, atLineNumber - 1);
-    }
-
-    /**
-     * If the inline diff does not work, we fall back to showing the diff on a paragraph base, i.e. deleting the old
-     * paragraph (adding the "deleted"-class) and adding the new one (adding the "added" class).
-     * If the provided Text is not wrapped in HTML elements but inline text, the returned text is using
-     * <ins>/<del>-tags instead of adding CSS-classes to the wrapping element.
-     *
-     * @param {string} oldText
-     * @param {string} newText
-     * @param {number|null} lineLength
-     * @param {number|null} firstLineNumber
-     * @returns {string}
-     */
-    private diffParagraphs(oldText: string, newText: string, lineLength: number, firstLineNumber: number): string {
-        let oldTextWithBreaks: Element;
-        let currChild: Element;
-
-        if (lineLength !== null) {
-            oldTextWithBreaks = this.lineNumberingService.insertLineNumbersNode(
-                oldText,
-                lineLength,
-                null,
-                firstLineNumber
-            );
-            newText = this.lineNumberingService.insertLineBreaksWithoutNumbers(newText, lineLength);
-        } else {
-            oldTextWithBreaks = document.createElement(`div`);
-            oldTextWithBreaks.innerHTML = oldText;
-        }
-        newText = newText.replace(/^\s+/g, ``).replace(/\s+$/g, ``);
-        const newTextWithBreaks = document.createElement(`div`);
-        newTextWithBreaks.innerHTML = newText;
-
-        for (const child of oldTextWithBreaks.childNodes) {
-            currChild = child as Element;
-            if (currChild.nodeType === TEXT_NODE) {
-                const wrapDel = document.createElement(`del`);
-                oldTextWithBreaks.insertBefore(wrapDel, currChild);
-                oldTextWithBreaks.removeChild(currChild);
-                wrapDel.appendChild(currChild);
-            } else {
-                DomHelpers.addCSSClass(currChild, `delete`);
-                this.removeColorStyles(currChild);
-            }
-        }
-        for (const child of newTextWithBreaks.childNodes) {
-            currChild = child as Element;
-            if (currChild.nodeType === TEXT_NODE) {
-                const wrapIns = document.createElement(`ins`);
-                newTextWithBreaks.insertBefore(wrapIns, currChild);
-                newTextWithBreaks.removeChild(currChild);
-                wrapIns.appendChild(currChild);
-            } else {
-                DomHelpers.addCSSClass(currChild, `insert`);
-                this.removeColorStyles(currChild);
-            }
-        }
-
-        const mergedFragment = document.createDocumentFragment();
-        let el: ChildNode;
-        while ((el = oldTextWithBreaks.firstChild)) {
-            oldTextWithBreaks.removeChild(el);
-            mergedFragment.appendChild(el);
-        }
-        while ((el = newTextWithBreaks.firstChild)) {
-            newTextWithBreaks.removeChild(el);
-            mergedFragment.appendChild(el);
-        }
-
-        return this.serializeDom(mergedFragment, false);
+        return null;
     }
 
     /**
@@ -1298,549 +205,19 @@ export class MotionDiffService {
         lineLength: number | null = null,
         firstLineNumber: number | null = null
     ): string {
-        const cacheKey = lineLength + ` ` + firstLineNumber + ` ` + djb2hash(htmlOld) + djb2hash(htmlNew);
-        const cached = this.diffCache.get(cacheKey);
-        if (cached) {
-            return cached;
-        }
-
-        // TODO: This is a workaround to make sure the first element of a amendment
-        //       has a line number for correct display of amendments in front of list
-        //       or block elements
-        const htmlOldEl = document.createElement(`template`);
-        const htmlNewEl = document.createElement(`template`);
-        htmlNewEl.innerHTML = htmlNew;
-        htmlOldEl.innerHTML = htmlOld;
-        if (htmlNewEl.content.children[0] && !htmlNewEl.content.children[0].querySelector(`.os-line-number`)) {
-            if (
-                htmlNewEl.content.querySelector(`.os-line-number`) &&
-                htmlOldEl.content.querySelector(`.os-line-number`)
-            ) {
-                const ln = htmlNewEl.content.querySelector(`.os-line-number`);
-                if (htmlNewEl.content.children[0].childNodes.length > 0) {
-                    htmlNewEl.content.children[0].childNodes[0].before(ln);
-                } else {
-                    htmlNewEl.content.children[0].before(ln);
-                }
-
-                htmlOldEl.content.children[0].querySelector(`.os-line-number`).remove();
-
-                htmlNew = htmlNewEl.innerHTML;
-                htmlOld = htmlOldEl.innerHTML;
-            }
-        }
-
-        const origHtmlNew = htmlNew;
-        const origHtmlOld = htmlOld;
-
-        // os-split-after should not be considered for detecting changes in paragraphs, so we strip it here
-        // and add it afterwards.
-        // We only do this for P for now, as for more complex types like UL/LI that tend to be nestend,
-        // information would get lost by this that we will need to recursively merge it again later on.
-        let isSplitAfter = false;
-        let isSplitBefore = false;
-        htmlOld = htmlOld.replace(
-            /(\s*<(?:p|ul|ol|li|blockquote|div)[^>]+class\s*=\s*["'][^"']*)os-split-after */gi,
-            (_match: string, beginning: string): string => {
-                isSplitAfter = true;
-                return beginning;
-            }
-        );
-        htmlNew = htmlNew.replace(
-            /(\s*<(?:p|ul|ol|li|blockquote|div)[^>]+class\s*=\s*["'][^"']*)os-split-after */gi,
-            (_match: string, beginning: string): string => {
-                isSplitAfter = true;
-                return beginning;
-            }
-        );
-        htmlOld = htmlOld.replace(
-            /(\s*<(?:p|ul|ol|li|blockquote|div)[^>]+class\s*=\s*["'][^"']*)os-split-before */gi,
-            (_match: string, beginning: string): string => {
-                isSplitBefore = true;
-                return beginning;
-            }
-        );
-        htmlNew = htmlNew.replace(
-            /(\s*<(?:p|ul|ol|li|blockquote|div)[^>]+class\s*=\s*["'][^"']*)os-split-before */gi,
-            (_match: string, beginning: string): string => {
-                isSplitBefore = true;
-                return beginning;
-            }
-        );
-
-        // Performing the actual diff
-        const str = this.diffString(htmlOld, htmlNew);
-        let diffUnnormalized = str.replace(/^\s+/g, ``).replace(/\s+$/g, ``).replace(/ {2,}/g, ` `);
-
-        diffUnnormalized = this.fixWrongChangeDetection(diffUnnormalized);
-
-        // Handles insertions in empty paragraphs
-        diffUnnormalized = diffUnnormalized.replace(
-            /<del>(<SPAN[^>]+os-line-number[^>]+?>)<\/del>(<ins>[\s\S]*?<\/ins>)\s<del><\/SPAN><\/del>/gi,
-            (_whole: string, span: string, insertedText: string): string =>
-                `<del>` + span + ` </SPAN></del>` + insertedText + `<ins> </ins>`
-        );
-
-        // Remove <del> tags that only delete line numbers
-        // We need to do this before removing </del><del> as done in one of the next statements
-        diffUnnormalized = diffUnnormalized.replace(
-            /<del>(((<BR CLASS="os-line-break">)<\/del><del>)?(<span[^>]+os-line-number[^>]+?>)(\s|<\/?del>)*<\/span>)<\/del>/gi,
-            (_found: string, _tag: string, _brWithDel: string, plainBr: string, span: string): string =>
-                (plainBr !== undefined ? plainBr : ``) + span + ` </span>`
-        );
-
-        // Merging individual insert/delete statements into bigger blocks
-        diffUnnormalized = diffUnnormalized.replace(/<\/ins><ins>/gi, ``).replace(/<\/del><del>/gi, ``);
-
-        // If we have a <del>deleted word</del>LINEBREAK<ins>new word</ins>, let's assume that the insertion
-        // was actually done in the same line as the deletion.
-        // We don't have the LINEBREAK-markers in the new string, hence we can't be a 100% sure, but
-        // this will probably the more frequent case.
-        // This only really makes a differences for change recommendations anyway, where we split the text into lines
-        // Hint: if there is no deletion before the line break, we have the same issue, but cannot solve this here.
-        diffUnnormalized = diffUnnormalized.replace(
-            /(<\/del>)(<BR CLASS="os-line-break"><span[^>]+os-line-number[^>]+?>\s*<\/span>)(<ins>[\s\S]*?<\/ins>)/gi,
-            (_found: string, del: string, br: string, ins: string): string => del + ins + br
-        );
-
-        // If only a few characters of a word have changed, don't display this as a replacement of the whole word,
-        // but only of these specific characters
-        diffUnnormalized = diffUnnormalized.replace(
-            /<del>([a-z0-9,_-]* ?)<\/del><ins>([a-z0-9,_-]* ?)<\/ins>/gi,
-            (_found: string, oldText: string, newText: string): string => {
-                let foundDiff = false;
-                let commonStart = ``;
-                let commonEnd = ``;
-                let remainderOld = oldText;
-                let remainderNew = newText;
-
-                while (remainderOld.length > 0 && remainderNew.length > 0 && !foundDiff) {
-                    if (remainderOld[0] === remainderNew[0]) {
-                        commonStart += remainderOld[0];
-                        remainderOld = remainderOld.substr(1);
-                        remainderNew = remainderNew.substr(1);
-                    } else {
-                        foundDiff = true;
-                    }
-                }
-
-                foundDiff = false;
-                while (remainderOld.length > 0 && remainderNew.length > 0 && !foundDiff) {
-                    if (remainderOld[remainderOld.length - 1] === remainderNew[remainderNew.length - 1]) {
-                        commonEnd = remainderOld[remainderOld.length - 1] + commonEnd;
-                        remainderNew = remainderNew.substr(0, remainderNew.length - 1);
-                        remainderOld = remainderOld.substr(0, remainderOld.length - 1);
-                    } else {
-                        foundDiff = true;
-                    }
-                }
-
-                let out = commonStart;
-                if (remainderOld !== ``) {
-                    out += `<del>` + remainderOld + `</del>`;
-                }
-                if (remainderNew !== ``) {
-                    out += `<ins>` + remainderNew + `</ins>`;
-                }
-                out += commonEnd;
-
-                return out;
-            }
-        );
-
-        // Replace spaces in line numbers by &nbsp;
-        diffUnnormalized = diffUnnormalized.replace(
-            /<span[^>]+os-line-number[^>]+?>\s*<\/span>/gi,
-            (found: string): string => found.toLowerCase().replace(/> <\/span/gi, `>&nbsp;</span`)
-        );
-
-        // The diff algorithm handles insertions in empty paragraphs as inserted in the next one
-        // <del><\/P><P><\/del><span>&nbsp;<\/span><ins>NEUER TEXT<\/P><P><\/ins>
-        // -> <ins>NEUER TEXT<\/ins><\/P><P><span>&nbsp;<\/span>
-        diffUnnormalized = diffUnnormalized.replace(
-            /<del>(<\/P><P>)<\/del>(<span[^>]+>&nbsp;<\/span>)(<del> <\/del>)?<ins>([\s\S]*?)\1<\/ins>/gi,
-            (_found: string, paragraph: string, span: string, _emptyDel: string, insText: string): string => {
-                return `<ins>` + insText + `</ins>` + paragraph + span;
-            }
-        );
-
-        // <P><ins>NEUE ZEILE</P>\n<P></ins> => <ins><P>NEUE ZEILE</P>\n</ins><P>
-        diffUnnormalized = diffUnnormalized.replace(
-            /<(p|div|blockquote|li)([^>]*)><(ins|del)>([\s\S]*?)<\/\1>(\s*)<(p|div|blockquote|li)([^>]*)><\/\3>/gi,
-            (
-                _whole: string,
-                block1: string,
-                att1: string,
-                insDel: string,
-                content: string,
-                space: string,
-                block2: string,
-                att2: string
-            ): string =>
-                `<` +
-                insDel +
-                `><` +
-                block1 +
-                att1 +
-                `>` +
-                content +
-                `</` +
-                block1 +
-                `>` +
-                space +
-                `</` +
-                insDel +
-                `><` +
-                block2 +
-                att2 +
-                `>`
-        );
-
-        // If larger inserted HTML text contains block elements, we separate the inserted text into
-        // inline <ins> elements and "insert"-class-based block elements.
-        // <ins>...<div>...</div>...</ins> => <ins>...</ins><div class="insert">...</div><ins>...</ins>
-        diffUnnormalized = diffUnnormalized.replace(
-            /<(ins|del)>([\s\S]*?)<\/\1>/gi,
-            (whole: string, insDel: string): string => {
-                const modificationClass = insDel.toLowerCase() === `ins` ? `insert` : `delete`;
-                return whole.replace(
-                    /(<(p|div|blockquote|ul|ol|li)[^>]*>)([\s\S]*?)(<\/\2>)/gi,
-                    (_whole2: string, opening: string, _blockTag: string, content: string, closing: string): string => {
-                        const modifiedTag = DomHelpers.addClassToHtmlTag(opening, modificationClass);
-                        return `</` + insDel + `>` + modifiedTag + content + closing + `<` + insDel + `>`;
-                    }
-                );
-            }
-        );
-
-        // <del>deleted text</P></del><ins>inserted.</P></ins> => <del>deleted text</del><ins>inserted.</ins></P>
-        diffUnnormalized = diffUnnormalized.replace(
-            /<del>([^<]*)<\/(p|div|blockquote|li)><\/del><ins>([^<]*)<\/\2>(\s*)<\/ins>/gi,
-            (_whole: string, deleted: string, tag: string, inserted: string, white: string): string =>
-                `<del>` + deleted + `</del><ins>` + inserted + `</ins></` + tag + `>` + white
-        );
-
-        // <ins>...</p><p>...</ins> => <ins>...</ins></p><p><ins>...</ins>
-        diffUnnormalized = diffUnnormalized.replace(
-            /<(ins|del)>([\s\S]*?)<\/(p|div|blockquote|li)>\s*<(p|div|blockquote|li)([^>]*)>([\s\S]*?)<\/\1>/gi,
-            (
-                whole: string,
-                insDel: string,
-                content1: string,
-                blockEnd: string,
-                blockStart: string,
-                blockAttrs: string,
-                content2: string
-            ): string => {
-                if (DomHelpers.isValidInlineHtml(content1) && DomHelpers.isValidInlineHtml(content2)) {
-                    return (
-                        `<` +
-                        insDel +
-                        `>` +
-                        content1 +
-                        `</` +
-                        insDel +
-                        `></` +
-                        blockEnd +
-                        `>` +
-                        `<` +
-                        blockStart +
-                        blockAttrs +
-                        `><` +
-                        insDel +
-                        `>` +
-                        content2 +
-                        `</` +
-                        insDel +
-                        `>`
-                    );
-                } else {
-                    return whole;
-                }
-            }
-        );
-
-        // Cleanup leftovers from the operation above, when <ins></ins>-tags ore <ins> </ins>-tags are left
-        // around block tags. It should be safe to remove them and just leave the whitespaces.
-        diffUnnormalized = diffUnnormalized.replace(
-            /<(ins|del)>(\s*)<\/\1>/gi,
-            (_whole: string, _insDel: string, space: string): string => space
-        );
-
-        // <del></p><ins> Added text</p></ins> -> <ins> Added text</ins></p>
-        diffUnnormalized = diffUnnormalized.replace(
-            /<del><\/(p|div|blockquote|li)><\/del><ins>([\s\S]*?)<\/\1>(\s*)<\/ins>/gi,
-            (_whole: string, blockTag: string, content: string, space: string): string =>
-                `<ins>` + content + `</ins></` + blockTag + `>` + space
-        );
-
-        // <ins><STRONG></ins>formatted<ins></STRONG></ins> => <del>formatted</del><ins><STRONG>formatted</STRONG></ins>
-        diffUnnormalized = diffUnnormalized.replace(
-            /<ins><(mark|span|strong|em|b|i|u|s|a|small|big|sup|sub)( [^>]*)?><\/ins>([^<]*)<ins><\/\1><\/ins>/gi,
-            (_whole: string, inlineTag: string, tagAttributes: string, content: string): string =>
-                `<del>` +
-                content +
-                `</del>` +
-                `<ins><` +
-                inlineTag +
-                (tagAttributes ? tagAttributes : ``) +
-                `>` +
-                content +
-                `</` +
-                inlineTag +
-                `></ins>`
-        );
-
-        // <del><STRONG></del>formatted<del></STRONG></del> => <del><STRONG>formatted</STRONG></del><ins>formatted</ins>
-        diffUnnormalized = diffUnnormalized.replace(
-            /<del><(mark|span|strong|em|b|i|u|s|a|small|big|sup|sub)( [^>]*)?><\/del>([^<]*)<del><\/\1><\/del>/gi,
-            (_whole: string, inlineTag: string, tagAttributes: string, content: string): string =>
-                `<del><` +
-                inlineTag +
-                (tagAttributes ? tagAttributes : ``) +
-                `>` +
-                content +
-                `</` +
-                inlineTag +
-                `></del>` +
-                `<ins>` +
-                content +
-                `</ins>`
-        );
-
-        // <del><STRONG>Körper</del><ins>alten <STRONG>Körpergehülle</ins></STRONG> => <ins>alten </ins><STRONG><del>Körper</del><ins>Körpergehülle</ins></STRONG>
-        diffUnnormalized = diffUnnormalized.replace(
-            /<del><(mark|span|strong|em|b|i|u|s|a|small|big|sup|sub)( [^>]*)?>([^<]*)<\/del><ins>([^<]*)<\1>([^<]*)<\/ins><\/\1>/gi,
-            (
-                _whole: string,
-                inlineTag: string,
-                tagAttributes: string,
-                delContent: string,
-                insContent1: string,
-                insContent2: string
-            ): string =>
-                `<ins>` +
-                insContent1 +
-                `</ins><` +
-                inlineTag +
-                (tagAttributes ? tagAttributes : ``) +
-                `><del>` +
-                delContent +
-                `</del>` +
-                `<ins>` +
-                insContent2 +
-                `</ins></` +
-                inlineTag +
-                `>`
-        );
-
-        // <del>with a </del><ins>a <STRONG></ins>unformatted <del>word</del><ins>sentence</STRONG></ins> ->
-        // <del>unformatted word</del><ins><STRONG>formatted word</STRONG></ins>
-        diffUnnormalized = diffUnnormalized.replace(
-            /<del>([^<]*)<\/del><ins><(mark|span|strong|em|b|i|u|s|a|small|big|sup|sub)( [^>]*)?>([^<]*)<\/ins>([^<]*)<ins><\/\2><\/ins>/gi,
-            (
-                _whole: string,
-                delContent: string,
-                inlineTag: string,
-                tagAttributes: string,
-                insContent: string,
-                unchangedContent: string
-            ): string =>
-                `<del>` +
-                delContent +
-                unchangedContent +
-                `</del><ins><` +
-                inlineTag +
-                (tagAttributes ? tagAttributes : ``) +
-                `>` +
-                insContent +
-                unchangedContent +
-                `</` +
-                inlineTag +
-                `></ins>`
-        );
-
-        // <ins><STRONG></ins>unformatted <del>word</del><ins>sentence</STRONG></ins> ->
-        // <del>unformatted word</del><ins><STRONG>unformatted sentence</STRONG></ins>
-        diffUnnormalized = diffUnnormalized.replace(
-            /<ins><(mark|span|strong|em|b|i|u|s|a|small|big|sup|sub)( [^>]*)?><\/ins>([^<]*)<del>([^<]*)<\/del><ins>([^<]*)<\/\1><\/ins>/gi,
-            (
-                _whole: string,
-                inlineTag: string,
-                tagAttributes: string,
-                unchangedContent: string,
-                delContent: string,
-                insContent: string
-            ): string =>
-                `<del>` +
-                unchangedContent +
-                delContent +
-                `</del><ins><` +
-                inlineTag +
-                (tagAttributes ? tagAttributes : ``) +
-                `>` +
-                unchangedContent +
-                insContent +
-                `</` +
-                inlineTag +
-                `></ins>`
-        );
-
-        // <STRONG>Bestätigung<del></STRONG></del><ins> NEU</STRONG> NEU2</ins> -->
-        // <STRONG>Bestätigung<ins> NEU</ins></STRONG><ins> NEU2</ins>
-        diffUnnormalized = diffUnnormalized.replace(
-            /<del>(<\/(mark|span|strong|em|b|i|u|s|a|small|big|sup|sub)>)<\/del><ins>([^>]*)<\/\2>([^>]*)<\/ins>/gi,
-            (
-                _whole: string,
-                closingTag: string,
-                closingTagInner: string,
-                insertedText1: string,
-                insertedText2: string
-            ): string => `<ins>` + insertedText1 + `</ins>` + closingTag + `<ins>` + insertedText2 + `</ins>`
-        );
-
-        // <del>Ebene 3 <UL><LI></del><span class="line-number-4 os-line-number" contenteditable="false" data-line-number="4">&nbsp;</span><ins>Ebene 3a <UL><LI></ins>
-        // => <del>Ebene 3 </del><ins>Ebene 3a </ins><UL><LI><span class="line-number-4 os-line-number" contenteditable="false" data-line-number="4">&nbsp;</span>
-        diffUnnormalized = diffUnnormalized.replace(
-            /<del>([^<]+)((?:<(?:ul|ol|li)>)+)<\/del>(<span[^>]*os-line-number[^>]*>(?:&nbsp;|\s)<\/span>)?<ins>([^<]+)\2<\/ins>/gi,
-            (_whole: string, del: string, block: string, ln: string, ins: string): string =>
-                `<del>` + del + `</del><ins>` + ins + `</ins>` + block + ln
-        );
-
-        // </p> </ins> -> </ins></p>
-        diffUnnormalized = diffUnnormalized.replace(
-            /(<\/(p|div|blockquote|li)>)(\s*)<\/(ins|del)>/gi,
-            (_whole: string, ending: string, _blockTag: string, space: string, insdel: string): string =>
-                `</` + insdel + `>` + ending + space
-        );
-
-        // <ul><li><ul><li>...</li><del></UL></LI></UL></del><LI class="insert">d</LI><LI class="insert">e</LI><ins></UL></LI></UL></ins>
-        // => <ul><li><ul><li>...</li><LI class="insert">d</LI><LI class="insert">e</LI></UL></LI></UL>
-        diffUnnormalized = diffUnnormalized.replace(
-            /<del>((<\/(li|ul|ol)>)+)<\/del>((<li class="insert">.*?<\/li>)*)<ins>\1<\/ins>/i,
-            (_whole: string, ending: string, _e1: string, _e2: string, insertedLis: string, _e3: string) =>
-                insertedLis + ending
-        );
-
-        // <UL><LI><UL><LI><UL><LI><del>Ebene 4</LI></UL></LI></UL></LI></UL></del><ins>Ebene 5</LI></UL></LI></UL></LI></UL></ins>
-        // => <UL><LI><UL><LI><UL><LI><del>Ebene 4</del><ins>Ebene 5</ins></LI></UL></LI></UL></LI></UL>
-        diffUnnormalized = diffUnnormalized.replace(
-            /<del>([^<>]*)((<\/(li|ul|ol)>)+)<\/del><ins>([^<>]*)\2<\/ins>/i,
-            (_whole, del, end, _ul, _u2, ins) => `<del>` + del + `</del><ins>` + ins + `</ins>` + end
-        );
-
-        let diff: string;
-        if (this.diffDetectBrokenDiffHtml(diffUnnormalized)) {
-            diff = this.diffParagraphs(htmlOld, htmlNew, lineLength!, firstLineNumber!);
-        } else {
-            let node: Element = document.createElement(`div`);
-            node.innerHTML = diffUnnormalized;
-            diff = node.innerHTML;
-
-            if (lineLength !== null && firstLineNumber !== null) {
-                node = this.lineNumberingService.insertLineNumbersNode(diff, lineLength, null, firstLineNumber);
-                diff = node.innerHTML;
-            }
-        }
-
-        if (isSplitAfter) {
-            diff = this.readdOsSplit(diff, [origHtmlOld, origHtmlNew]);
-        }
-        if (isSplitBefore) {
-            diff = this.readdOsSplit(diff, [origHtmlOld, origHtmlNew], true);
-        }
-
-        this.diffCache.put(cacheKey, diff);
-        return diff;
+        return null;
     }
 
     public readdOsSplit(diff: string, versions: string[], before = false): string {
-        const className = before ? `os-split-before` : `os-split-after`;
-
-        const diffEl = document.createElement(`template`);
-        diffEl.innerHTML = diff;
-        const diffNode = (before ? diffEl.content.firstChild : diffEl.content.lastChild) as HTMLElement;
-
-        const versionNodes: HTMLElement[] = [];
-        let found = false;
-        for (const v of versions) {
-            const el = document.createElement(`template`);
-            el.innerHTML = v;
-            versionNodes.push((before ? el.content.firstChild : el.content.lastChild) as HTMLElement);
-            found = found || !!el.content.querySelector(`.${className}`);
-        }
-
-        if (!found) {
-            return diff;
-        }
-
-        this.recAddOsSplit(diffNode, versionNodes, before);
-
-        return diffEl.innerHTML;
-    }
-
-    private recAddOsSplit(diff: HTMLElement, versions: HTMLElement[], before = false): void {
-        const className = before ? `os-split-before` : `os-split-after`;
-        let containsSplit = false;
-        for (const v of versions) {
-            if (v?.classList.contains(className)) {
-                containsSplit = true;
-            }
-        }
-
-        if (!containsSplit) {
-            return;
-        }
-
-        const nextVersions: HTMLElement[] = [];
-        for (const v of versions) {
-            const s = v?.querySelector(`& > .${className}`) as HTMLElement;
-            if (s) {
-                nextVersions.push(s);
-            }
-        }
-
-        diff?.classList?.add(className);
-        const nextDiffNode = diff?.querySelector(`& > *:not(.os-line-number)`) as HTMLElement;
-        if (nextDiffNode) {
-            this.recAddOsSplit(nextDiffNode, nextVersions, before);
-        }
-
-        if (
-            diff.nextElementSibling &&
-            ((diff.classList.contains(`delete`) && diff.nextElementSibling.classList.contains(`insert`)) ||
-                (diff.classList.contains(`insert`) && diff.nextElementSibling.classList.contains(`delete`)))
-        ) {
-            diff.nextElementSibling.classList?.add(className);
-            const nextSibDiffNode = diff.nextElementSibling.querySelector(`& > *:not(.os-line-number)`) as HTMLElement;
-            if (nextSibDiffNode) {
-                this.recAddOsSplit(nextSibDiffNode, nextVersions, before);
-            }
-        }
+        return null;
     }
 
     public changeHasCollissions(change: ViewUnifiedChange, changes: ViewUnifiedChange[]): boolean {
-        return (
-            changes.filter(
-                (otherChange: ViewUnifiedChange) =>
-                    otherChange.getChangeId() !== change.getChangeId() &&
-                    ((otherChange.getLineFrom() >= change.getLineFrom() &&
-                        otherChange.getLineFrom() <= change.getLineTo()) ||
-                        (otherChange.getLineTo() >= change.getLineFrom() &&
-                            otherChange.getLineTo() <= change.getLineTo()) ||
-                            (otherChange.getLineFrom() <= change.getLineFrom() &&
-                                otherChange.getLineTo() >= change.getLineTo()))
-            ).length > 0
-        );
+        return null;
     }
 
     public sortChangeRequests(changes: ViewUnifiedChange[]): ViewUnifiedChange[] {
-        return changes.sort((change1: ViewUnifiedChange, change2: ViewUnifiedChange): number => {
-            if (change1.getLineFrom() === change2.getLineFrom()) {
-                return change1.getIdentifier() < change2.getIdentifier() ? -1 : 1;
-            }
-            return change1.getLineFrom() - change2.getLineFrom();
-        });
+        return null;
     }
 
     /**
@@ -1861,89 +238,14 @@ export class MotionDiffService {
         highlightLine?: number,
         firstLine = 1
     ): string {
-        let html = motionHtml;
-
-        changes = changes.filter(change => !change.isTitleChange());
-        // Changes need to be applied from the bottom up, to prevent conflicts with changing line numbers.
-        changes = this.sortChangeRequests(changes).reverse();
-
-        if (showAllCollisions) {
-            let lastReplacedLine: number = null;
-
-            changes.forEach(change => {
-                html = this.lineNumberingService.insertLineNumbers({ html, lineLength, firstLine: firstLine });
-
-                if (this.changeHasCollissions(change, changes)) {
-                    // In case of colliding amendments, we remove the original text first before inserting the amendments one by one.
-                    // Note: if amendment 1 affects line 3-5, we remove 3-5. If amendment 2 affects line 2-4, we only need to remove
-                    // line 2, as 3-5 is already removed. If Amendment 3 affects 2-4 too, we don't have to remove anything anymore.
-
-                    let removeUntil = change.getLineTo();
-                    if (lastReplacedLine !== null && lastReplacedLine <= removeUntil) {
-                        removeUntil = lastReplacedLine - 1;
-                    }
-                    if (removeUntil >= change.getLineFrom()) {
-                        html = this.removeLines(html, change.getLineFrom(), removeUntil);
-                        html = this.lineNumberingService.insertLineNumbers({ html, lineLength, firstLine: firstLine });
-                    }
-
-                    const type =
-                        ` data-change-type="` +
-                        ((type: ViewUnifiedChangeType): `amendment` | `recommendation` | `unknown` => {
-                            switch (type) {
-                                case ViewUnifiedChangeType.TYPE_AMENDMENT:
-                                    return `amendment`;
-                                case ViewUnifiedChangeType.TYPE_CHANGE_RECOMMENDATION:
-                                    return `recommendation`;
-                                default:
-                                    return `unknown`;
-                            }
-                        })(change.getChangeType()) +
-                        `"`;
-                    const changeId = ` data-change-id="` + DomHelpers.replaceHtmlEntities(change.getChangeId()) + `"`;
-                    const title = ` data-title="` + DomHelpers.replaceHtmlEntities(change.getTitle()) + `"`;
-                    const ident = ` data-identifier="` + DomHelpers.replaceHtmlEntities(change.getIdentifier()) + `"`;
-                    const lineFrom = ` data-line-from="` + change.getLineFrom().toString(10) + `"`;
-                    const lineTo = ` data-line-to="` + change.getLineTo().toString(10) + `"`;
-                    const opAttrs = type + ident + title + changeId + lineFrom + lineTo;
-                    const opTag = `<div class="os-colliding-change os-colliding-change-holder"` + opAttrs + `>`;
-                    const insertingHtml = opTag + change.getChangeNewText() + `</div>`;
-
-                    html = this.insertLines(html, change.getLineFrom(), insertingHtml);
-
-                    lastReplacedLine = change.getLineFrom();
-                } else {
-                    html = this.replaceLines(html, change.getChangeNewText(), change.getLineFrom(), change.getLineTo());
-                }
-            });
-        } else {
-            changes.forEach((change: ViewUnifiedChange) => {
-                html = this.lineNumberingService.insertLineNumbers({ html, lineLength, firstLine: firstLine });
-                html = this.replaceLines(html, change.getChangeNewText(), change.getLineFrom(), change.getLineTo());
-            });
-        }
-
-        html = this.lineNumberingService.insertLineNumbers({
-            html,
-            lineLength,
-            highlight: highlightLine,
-            firstLine: firstLine
-        });
-
-        return html;
+        return null;
     }
 
     public formatOsCollidingChanges(
         html: string,
         formatter: (el: HTMLDivElement, type: string, identifier: string, title: string, changeId: string) => void
     ): string {
-        const frag = DomHelpers.htmlToFragment(html);
-
-        frag.querySelectorAll(`.os-colliding-change`).forEach((el: HTMLElement): void => {
-            formatter.bind(this)(el as HTMLDivElement);
-        });
-
-        return DomHelpers.fragmentToHtml(frag);
+        return null;
     }
 
     public formatOsCollidingChanges_wysiwyg_cb(el: HTMLDivElement): void {
@@ -2000,7 +302,7 @@ export class MotionDiffService {
         }
 
         // Change recommendations do not have a title
-        let strTitle;
+        let strTitle: string;
         if (type === `recommendation`) {
             strTitle = this.translate.instant(`Change recommendation`);
         } else {
@@ -2015,7 +317,7 @@ export class MotionDiffService {
         const comment = el.ownerDocument.createElement(commentsInInlineElement ? `span` : `div`);
         comment.classList.add(`os-colliding-change`);
         comment.classList.add(`os-colliding-change-comment`);
-        comment.innerHTML = `&lt;` + DomHelpers.replaceHtmlEntities(`!-- ### ` + strTitle + ` ### --`) + `&gt;`;
+        comment.innerHTML = `&lt;` + replaceHtmlEntities(`!-- ### ` + strTitle + ` ### --`) + `&gt;`;
         if (commentsInInlineElement) {
             comment.innerHTML = comment.innerHTML + `<br>`;
         }
@@ -2053,56 +355,7 @@ export class MotionDiffService {
         lineLength: number,
         changeRecos?: ViewUnifiedChange[]
     ): DiffLinesInParagraph | null {
-        const paragraph_line_range: LineNumberRange = this.lineNumberingService.getLineNumberRange(origText);
-        let diff = this.diff(origText, newText);
-        const affected_lines = this.detectAffectedLineRange(diff) as LineRange;
-
-        /**
-         * If the affect line has change recos, overwirte the diff with the change reco
-         */
-        if (changeRecos && changeRecos.length) {
-            const recoToThisLine = changeRecos.find(reco => reco.getLineFrom() === affected_lines.from);
-            if (recoToThisLine) {
-                diff = this.diff(origText, recoToThisLine.getChangeNewText());
-            }
-        }
-
-        if (affected_lines === null) {
-            return null;
-        }
-
-        let textPre = ``;
-        let textPost = ``;
-        if (affected_lines.from > paragraph_line_range.from!) {
-            textPre = this.formatDiffWithLineNumbers(
-                this.extractRangeByLineNumbers(diff, paragraph_line_range.from!, affected_lines.from - 1),
-                lineLength,
-                paragraph_line_range.from!
-            );
-        }
-        if (paragraph_line_range.to! > affected_lines.to) {
-            textPost = this.formatDiffWithLineNumbers(
-                this.extractRangeByLineNumbers(diff, affected_lines.to + 1, paragraph_line_range.to!),
-                lineLength,
-                affected_lines.to + 1
-            );
-        }
-        const text = this.formatDiffWithLineNumbers(
-            this.extractRangeByLineNumbers(diff, affected_lines.from, affected_lines.to),
-            lineLength,
-            affected_lines.from
-        );
-
-        return {
-            paragraphNo,
-            paragraphLineFrom: paragraph_line_range.from,
-            paragraphLineTo: paragraph_line_range.to,
-            diffLineFrom: affected_lines.from,
-            diffLineTo: affected_lines.to,
-            textPre,
-            text,
-            textPost
-        } as DiffLinesInParagraph;
+        return null;
     }
 
     /**
@@ -2121,41 +374,7 @@ export class MotionDiffService {
         lineLength: number,
         highlight?: number
     ): string {
-        if (this.lineNumberingService.getLineNumberRange(html).to < change.getLineTo()) {
-            throw new Error(`Invalid call - The change is outside of the motion`);
-        }
-        const data: ExtractedContent = this.extractRangeByLineNumbers(html, change.getLineFrom(), change.getLineTo());
-        let oldText =
-            data.outerContextStart +
-            data.innerContextStart +
-            data.html +
-            data.innerContextEnd +
-            data.outerContextEnd;
-
-        oldText = this.lineNumberingService.insertLineNumbers({
-            html: oldText,
-            lineLength,
-            firstLine: change.getLineFrom()
-        });
-        let diff = this.diff(oldText, change.getChangeNewText());
-
-        // If an insertion makes the line longer than the line length limit, we need two line breaking runs:
-        // - First, for the official line numbers, ignoring insertions (that's been done some lines before)
-        // - Second, another one to prevent the displayed including insertions to exceed the page width
-        diff = this.lineNumberingService.insertLineBreaksWithoutNumbers(diff, lineLength, true);
-
-        if (highlight && highlight > 0) {
-            diff = this.lineNumberingService.highlightLine(diff, highlight);
-        }
-
-        const origBeginning = data.outerContextStart + data.innerContextStart;
-        if (diff.toLowerCase().indexOf(origBeginning.toLowerCase()) === 0) {
-            // Add "merge-before"-css-class if the first line begins in the middle of a paragraph. Used for PDF.
-            diff =
-                DomHelpers.addCSSClassToFirstTag(origBeginning, `merge-before`) + diff.substring(origBeginning.length);
-        }
-
-        return diff;
+        return null;
     }
 
     /**
@@ -2175,42 +394,7 @@ export class MotionDiffService {
         highlight?: number,
         lineRange?: LineRange
     ): string {
-        if (changes.length === 0 && !lineRange) {
-            return motionHtml;
-        }
-
-        let maxFromLine = lineRange?.from || this.lineNumberingService.getLineNumberRange(motionHtml).from - 1;
-        const maxToLine = lineRange?.to || this.lineNumberingService.getLineNumberRange(motionHtml).to;
-        let hasRemainederOneChangedLine = false;
-
-        for (const change of changes) {
-            if (change.getLineTo() > maxFromLine && change.getLineTo() <= maxToLine) {
-                maxFromLine = change.getLineTo();
-                hasRemainederOneChangedLine = true;
-            }
-        };
-
-        if (!hasRemainederOneChangedLine) {
-            return ``;
-        }
-
-        const data: ExtractedContent = this.extractRangeByLineNumbers(
-            motionHtml,
-            Math.max(maxFromLine + 1, lineRange?.from || 1),
-            lineRange?.to ? maxToLine : null
-        );
-
-        let html = ``;
-        if (data.html !== ``) {
-            // Add "merge-before"-css-class if the first line begins in the middle of a paragraph. Used for PDF.
-            html =
-                DomHelpers.addCSSClassToFirstTag(data.outerContextStart + data.innerContextStart, `merge-before`) +
-                data.html +
-                data.innerContextEnd +
-                data.outerContextEnd;
-            html = this.lineNumberingService.insertLineNumbers({ html, lineLength, highlight, firstLine: maxFromLine + 1 });
-        }
-        return html;
+        return null;
     }
 
     /**
@@ -2229,22 +413,6 @@ export class MotionDiffService {
         lineLength: number,
         highlightedLine?: number
     ): string {
-        let html = ``;
-        const extracted = this.extractRangeByLineNumbers(motionText, lineRange.from, lineRange.to);
-        html =
-            extracted.outerContextStart +
-            extracted.innerContextStart +
-            extracted.html +
-            extracted.innerContextEnd +
-            extracted.outerContextEnd;
-        if (lineNumbers) {
-            html = this.lineNumberingService.insertLineNumbers({
-                html,
-                lineLength,
-                highlight: highlightedLine,
-                firstLine: lineRange.from
-            });
-        }
-        return html;
+        return null;
     }
 }
