+++ packages/openslides-motion-diff/karma.conf.ts
@@ -1,0 +1,30 @@
+process.env.TZ = 'UTC';
+
+// Karma configuration file, see link for more information
+// https://karma-runner.github.io/1.0/config/configuration-file.html
+module.exports = (config) => {
+  config.set({
+    basePath: '',
+    frameworks: ['jasmine'],
+    plugins: [
+      require('karma-jasmine'),
+      require('karma-jsdom-launcher')
+    ],
+    client: {
+      jasmine: {
+        // you can add configuration options for Jasmine here
+        // the possible options are listed at https://jasmine.github.io/api/edge/Configuration.html
+        // for example, you can disable the random execution with `random: false`
+        // or set a specific seed with `seed: 4321`
+      },
+      clearContext: false // leave Jasmine Spec Runner output visible in browser
+    },
+    jasmineHtmlReporter: {
+      suppressAll: true // removes the duplicated traces
+    },
+    browsers: ['jsdom'],
+    singleRun: false,
+    restartOnFileChange: true,
+    browserNoActivityTimeout: 400000
+  });
+};
